Fix nightMode in Card Type template preview
This was done by adding `.card` or `.card.nightMode` to the `:root` and
`:root.nightMode` selectors respectively.

The nightMode was working in the usual scenarios (reviewing card,
preview window, etc.). However, if one went to the menu where HTML and
CSS of the Card Type can be viewed and modified, then changing the
Options->Night Mode checkbox didn't do much for the card in the Preview
window, it always stayed in the dark mode.

`:root` selector could not be replaced completely with the `.card`
selector as otherwise theme preview in the browser would not work.