feat: add option for making highlighted code sideways scrollable
This commit adds a new "scrollable code blocks" option to the
config file. When set to True, the highlighted code blocks in
generated Anki cards will be sideways scrollable.

This option was made specifically for the mobile displays where
tall, narrow screen would cause the long code lines to wrap around
and interfere with the readability.

This option is by default disabled, to ensure that it's addition doesn't
break existing users.